Rewrite downloader
The downloader is now similar to the WOF downloader in that it uses
child_process.exec and curl to download data. This seems to be more
reliable than using request and piping to a file.

There were also issues with the progress bar package, so rather than
sort them out, this allows us to simply remove them.

This change probably will help fix https://github.com/pelias/geonames/issues/26,
but it's not 100% certain.