Explicitly handle metadata in Docker and update Documentation
The mystery of metadata from https://github.com/pelias/geonames/issues/222
has been solved! It turns out npm postinstall hooks, which we use to
download the metadata, do not fire when `DEBIAN_FRONTEND=noninteractive`
is set in the shell, such as we do in our
Docker [baseimage](https://github.com/pelias/dockerfiles/blob/master/baseimage/Dockerfile#L8).

This explains why the metadata downloads fine on local machines and in
the Chef Mapzen Search environment, but not in Kubernetes. It also
doesn't work in the Docker image in general.

This PR adds an explicit `npm run download_metadata` to the Dockerfile,
so that the Docker images contain this metadata, solving any Docker or
Kubernetes issues.

Additionally, it updates the readme to both clarify the situation
in the metadata documentation section, and remove it from the main
install instructions (since under normal interactive shell sessions it
is _not_ required and would just be an extra step).

Fixes https://github.com/pelias/geonames/issues/222
Connects https://github.com/pelias/geonames/pull/223